git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/sgopt@4343 4eea3252-f0fb-4393-894d-40516dce545b

+++ base/src/sgpp/base/grid/generation/GridGenerator.hpp
@@ -37,7 +37,7 @@ namespace SGPP {
          *
          * @param level Grid level
          */
-        virtual void regular(int level) = 0;
+        virtual void regular(size_t level) = 0;
 
 
         /**
@@ -46,19 +46,19 @@ namespace SGPP {
          * @param level Grid level
          * @param clique_size clique size
          */
-        virtual void cliques(int level, size_t clique_size) = 0;
+        virtual void cliques(size_t level, size_t clique_size) = 0;
 
         /**
          * Creates a full grid for a certain level @f$ n @f$, i.e., @f$ V_n = \bigoplus_{|\vec{l}|_\infty \leq n} W_{\vec{l}}@f$.
          *
          * @param level Grid level
          */
-        virtual void full(int level) = 0;
+        virtual void full(size_t level) = 0;
 
         /**
          * Creates a grid which doesn't contain the fullgrids with li<l_user, for any li level_t
          * */
-        virtual void truncated(int level, size_t l_user) {};
+        virtual void truncated(size_t level, size_t l_user) {};
 
         /**
          * Refines a grid according to the settings of the RefinementFunctor func.
@@ -106,7 +106,7 @@ namespace SGPP {
          * @param func pointer to refinement functor
          * @param maxLevel no points on higher levels than maxLevel will be created
          */
-        virtual void refineMaxLevel(RefinementFunctor* func, int maxLevel) = 0;
+        virtual void refineMaxLevel(RefinementFunctor* func, size_t maxLevel) = 0;
 
         /**
          * Returns the number of points on the grid that can be refined in the next iteration
@@ -116,7 +116,7 @@ namespace SGPP {
          *
          * @return the number of points on the grid that can be refined
          */
-        virtual size_t getNumberOfRefinablePointsToMaxLevel(int maxLevel) = 0;
+        virtual size_t getNumberOfRefinablePointsToMaxLevel(size_t maxLevel) = 0;
     };
 
   }
